requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
conda==4.3.16
Flask==2.0.2
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycosat==0.6.3
requests==2.27.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.7.0
mysql-connector==2.2.9
